requirements.txt
asgiref==3.8.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==4.2.20
django-allauth==65.14.3
django-axes==8.3.1
django-csp==4.0
django-debug-toolbar==6.2.0
django-jazzmin==3.0.1
django-ratelimit==4.1.0
idna==3.11
packaging==26.0
pillow==11.1.0
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.3
